Added impl layer for repositories
